feat: update according to comments
