+++ packages/core/src/web/app/stores/googleFontStore/types.ts
@@ -54,9 +54,9 @@ export interface GoogleFontActions {
   getFallbackPostScriptName: (fallbackFont: string) => string;
   getRegisteredFont: (postscriptName: string) => GoogleFont | undefined;
   isGoogleFontLoaded: (fontFamily: string) => boolean;
+  isLocalFont: (fontFamily: string) => boolean;
   isNetworkAvailableForGoogleFonts: () => boolean;
   isRegistered: (postscriptName: string) => boolean;
-  isWebSafeFont: (fontFamily: string) => boolean;
   loadGoogleFont: (fontFamily: string) => Promise<void>;
   loadGoogleFontBinary: (
     fontFamily: string,
