+++ packages/core/src/web/app/stores/googleFontStore/utils/detection.ts
@@ -2,16 +2,29 @@ import localFontHelper from '@core/implementations/localFontHelper';
 
 import { ICON_FONT_KEYWORDS, WEB_SAFE_FONTS } from '../constants';
 
+/**
+ * Normalize font family names for robust comparison
+ * Handles various font name formats and edge cases
+ */
+const normalizeFontName = (fontName: string): string =>
+  fontName
+    .toLowerCase()
+    .trim()
+    .replace(/^['"]+|['"]+$/g, '') // Remove surrounding quotes
+    .replace(/\s+/g, ' ') // Normalize whitespace
+    .replace(/[^\w\s-]/g, '') // Remove special characters except spaces and hyphens
+    .trim();
+
+const isWebSafeFont = (fontFamily: string): boolean => {
+  const normalizedFamily = fontFamily.toLowerCase().trim();
+
+  return WEB_SAFE_FONTS.some((webSafeFont) => normalizedFamily === webSafeFont.toLowerCase());
+};
+
 export const isIconFont = (fontFamily: string): boolean => {
   const lowerName = fontFamily.toLowerCase();
 
   return ICON_FONT_KEYWORDS.some((keyword) => lowerName.includes(keyword));
-};
-
-export const isWebSafeFont = (fontFamily: string): boolean => {
-  const normalizedFamily = fontFamily.toLowerCase().trim();
-
-  return WEB_SAFE_FONTS.some((webSafeFont) => normalizedFamily === webSafeFont.toLowerCase());
 };
 
 export const isLocalFont = (fontFamily: string): boolean => {
@@ -22,6 +35,15 @@ export const isLocalFont = (fontFamily: string): boolean => {
   }
 
   if (isWebSafeFont(fontFamily)) {
+    return true;
+  }
+
+  // Additional check: get all available local fonts and do normalized comparison
+  const localFonts = localFontHelper.getAvailableFonts();
+  const normalizedTarget = normalizeFontName(fontFamily);
+  const directMatch = localFonts.some((font) => normalizeFontName(font.family) === normalizedTarget);
+
+  if (directMatch) {
     return true;
   }
